Semi-responsive implementation
This is a very basic implementation right now. I haven’t been able to
really focus on this all week, but as soon as I import the assets over,
I’ll be able to focus on integrating responsiveness and distinct
stylesheets within the project. I’m also working on the horizontal
scrolling aspect, as my old workaround made it very counterintuitive.